+++ src/model/documentation/blocks/SDKDocumentationPageBlockFrame.ts
@@ -6,6 +6,8 @@
 //  Copyright © 2021 Supernova. All rights reserved.
 //
 
+import { ColorTokenRemoteData } from "model/tokens/remote/SDKRemoteTokenData"
+
 // --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- ---
 // MARK: - Imports
 
@@ -20,9 +22,7 @@ export interface DocumentationPageBlockFrameModel {
 
     title?: string
     description?: string
-    backgroundColor?: {
-      value?: string | null
-    }
+    color?: ColorTokenRemoteData
 
     origin: {
         sourceFileName: string
@@ -47,12 +47,12 @@ export class DocumentationPageBlockFrame {
     title: string | null
     description: string | null
     previewUrl: string
-    backgroundColor: string | null
+    backgroundColor: ColorTokenRemoteData | null
 
   // --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- ---
   // MARK: - Constructor
 
-  constructor(model: DocumentationPageBlockFrameModel, color: string | null) {
+  constructor(model: DocumentationPageBlockFrameModel, blockBackgroundColor: ColorTokenRemoteData | null) {
 
     this.sourceFileId = model.sourceFileId
     this.sourceFrameId = model.sourceFrameId
@@ -61,6 +61,6 @@ export class DocumentationPageBlockFrame {
     this.title = model.title.length > 0 ? model.title : model.origin.title ?? null
     this.description = model.description.length > 0 ? model.description : null
     this.previewUrl = model.origin.previewUrl
-    this.backgroundColor = model.backgroundColor?.value ?? color ?? null
+    this.backgroundColor = model.color ?? blockBackgroundColor ?? null
   }
 }