+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this SDK are mentioned here in this file.
 
+## [1.6.20] - 2022-18-02
+### Data Model Fixes
+- Fixed `backgroundColor` computation for `Frames` and `Frame`
+- Fixed `backgroundColor` computation for `Assets` and `Asset`
 
 ## [1.6.19] - 2022-18-02
 ### Data Model Extension